refactor(ui): colocate send guard message matching
