+++ ui/src/components/chat/ChatInterface.tsx
@@ -24,7 +24,7 @@ import { normalizeSessionTimestamps } from "@/lib/sessionTimestamps";
 import { getAgentWithResolvedKind, waitForSandboxAgentReady } from "@/app/actions/agents";
 import { toast } from "sonner";
 import { useRouter } from "next/navigation";
-import { createMessageHandlers, extractMessagesFromTasks, extractApprovalMessagesFromTasks, extractTokenStatsFromTasks, createMessage, ADKMetadata, ProcessedToolCallData } from "@/lib/messageHandlers";
+import { createMessageHandlers, extractMessagesFromTasks, extractApprovalMessagesFromTasks, extractTokenStatsFromTasks, createMessage, countSendGuardComparableMessages, countBackendBackedComparableMessages, ADKMetadata, ProcessedToolCallData } from "@/lib/messageHandlers";
 import { kagentA2AClient } from "@/lib/a2aClient";
 import { formatA2AClientError } from "@/lib/a2aErrors";
 import { useChatRunInSandbox, useChatSubstrateSandbox } from "@/components/chat/ChatAgentContext";
@@ -37,87 +37,6 @@ const RESUBSCRIBE_TASK_STATES: TaskState[] = ["submitted", "working"];
 // Task states that mean the session is busy (used by the cross-tab send guard).
 const ACTIVE_TASK_STATES: TaskState[] = ["submitted", "working", "input-required"];
 
-const SEND_GUARD_EXCLUDED_ORIGINAL_TYPES = new Set([
-  "ToolApprovalRequest", // HITL approval UI duplicates the pending backend task state.
-  "AskUserRequest", // ask_user UI duplicates the pending backend task state.
-  "ToolCallSummaryMessage", // UI-only marker that closes tool calls; not a backend chat turn.
-]);
-
-function isSendGuardComparableMessage(message: Message): boolean {
-  const meta = message.metadata as ADKMetadata | undefined;
-  return !meta?.originalType || !SEND_GUARD_EXCLUDED_ORIGINAL_TYPES.has(meta.originalType);
-}
-
-function countSendGuardComparableMessages(messages: Message[]): number {
-  let count = 0;
-  for (const message of messages) {
-    if (isSendGuardComparableMessage(message)) {
-      count += 1;
-    }
-  }
-  return count;
-}
-
-function getSendGuardMessageContentSignature(message: Message): string {
-  const meta = message.metadata as ADKMetadata | undefined;
-  const originalType = meta?.originalType === "TextMessage" ? "" : (meta?.originalType ?? "");
-  const partsSignature = message.parts
-    ?.map(part => part.kind === "text" ? `text:${part.text ?? ""}` : `${part.kind}:${JSON.stringify(part)}`)
-    .join("\u0001") ?? "";
-
-  return [
-    message.role,
-    originalType,
-    JSON.stringify(meta?.toolCallData ?? null),
-    JSON.stringify(meta?.toolResultData ?? null),
-    partsSignature,
-  ].join("\u0000");
-}
-
-function getSendGuardMessageKey(message: Message): string | undefined {
-  const contentSignature = getSendGuardMessageContentSignature(message);
-
-  if (message.role === "user" && message.messageId) {
-    return ["message", message.messageId].join("\u0000");
-  }
-
-  // Same-tab streamed agent display messages are locally re-created, so their
-  // messageId may differ from the backend history item. The task/context pair
-  // is the stable backend identity for those messages; the richer signature
-  // avoids counting unrelated messages in the same task.
-  if (message.contextId && message.taskId) {
-    return ["task", message.contextId, message.taskId, contentSignature].join("\u0000");
-  }
-
-  if (message.messageId) {
-    return ["message", message.messageId].join("\u0000");
-  }
-
-  return undefined;
-}
-
-function countBackendBackedComparableMessages(localMessages: Message[], backendMessages: Message[]): number {
-  const backendCounts = new Map<string, number>();
-  for (const message of backendMessages) {
-    if (!isSendGuardComparableMessage(message)) continue;
-    const key = getSendGuardMessageKey(message);
-    if (!key) continue;
-    backendCounts.set(key, (backendCounts.get(key) ?? 0) + 1);
-  }
-
-  let count = 0;
-  for (const message of localMessages) {
-    if (!isSendGuardComparableMessage(message)) continue;
-    const key = getSendGuardMessageKey(message);
-    if (!key) continue;
-    const remaining = backendCounts.get(key) ?? 0;
-    if (remaining > 0) {
-      count += 1;
-      backendCounts.set(key, remaining - 1);
-    }
-  }
-  return count;
-}
 
 interface ChatInterfaceProps {
   selectedAgentName: string;
